requirements.txt
pyqt5
pillow
torch
torchvision
pandas

